# Optimize partition_slice