@sunfishcode 2018-05-18
Nice! And TIL `DoubleEndedIterator` :-).